requirements.txt
Random-Word==1.0.11
aiogram==2.25.1
python-dotenv==1.0.0